requirements.txt
audioread==3.0.1
bidict==0.23.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
decorator==5.1.1
essentia-tensorflow==2.1b6.dev1177
Flask==3.1.0
Flask-SocketIO==5.5.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
MarkupSafe==3.0.2
msgpack==1.1.0
numba==0.61.0
numpy==1.26.4
packaging==24.2
platformdirs==4.3.6
pooch==1.8.2
PyAudio==0.2.14
pycparser==2.22
python-engineio==4.11.2
python-socketio==5.12.1
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.0
simple-websocket==1.1.0
six==1.17.0
sounddevice==0.5.1
soundfile==0.13.1
soxr==0.5.0.post1
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
